+++ pkg/backend/display/tableutil.go
@@ -1,0 +1,38 @@
+// Copyright 2016-2018, Pulumi Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package display
+
+import (
+	"strings"
+	"unicode/utf8"
+
+	"github.com/pulumi/pulumi/pkg/diag/colors"
+	"github.com/pulumi/pulumi/pkg/util/contract"
+)
+
+func columnHeader(msg string) string {
+	return colors.Underline + colors.BrightBlue + msg + colors.Reset
+}
+
+func messagePadding(uncolorizedColumn string, maxLength, extraPadding int) string {
+	extraWhitespace := maxLength - utf8.RuneCountInString(uncolorizedColumn)
+	contract.Assertf(extraWhitespace >= 0, "Neg whitespace. %v %s", maxLength, uncolorizedColumn)
+
+	// Place two spaces between all columns (except after the first column).  The first
+	// column already has a ": " so it doesn't need the extra space.
+	extraWhitespace += extraPadding
+
+	return strings.Repeat(" ", extraWhitespace)
+}